Update Conceptual Design.md
Changed the word proficient in that sentence, but I still don't know if I changed it to be what you are trying to say in that sentence @Evan

Added shall statements to most of the subsystem sections. (I did not touch the operating system/brain of robot section because I did not know the best way to formulate those into shall statements)(also did not touch drivetrain as this whole section will likely be changed whenever ME team sends us that kit they are getting)

Deleted the robot chassis section per Micah's request. 

Added a lidar camera/depth sensor subsystem. I left this blank for whoever knows what they are talking about. Confused on what to even put here because I feel like it will repeat what is already stated previously in the document but?

I put a note under cable laying device subsystem. I basically left this subsystem in just in case we decide we need a motor for the cable dispensing out of the box. 

I don't think that the division of labor should be changed, but I will let y'all look at it as well. Seems to be fine to list their strengths and then the subsystem they are in charge of based on that but idk.